# perf(gnovm): cache PkgIDFromPkgPath for higher performance