@odeke-em 2024-12-29
> Nice one, looks good for now IMO
> 
> The cache might use a little bit too much memory in the far future but we can revisit then
> 
> Please pass CI :)

@n0izn0iz I considered using a small slice say pre-allocated 100 items as the cache then linear looking up but that wasn't much different.
```diff
-var pkgIDFromPkgPathCache = make(map[string]*PkgID)
+type pkgCached struct {
+       path string
+       id   *PkgID
+}
+
+var pkgIDFromPkgPathCache = make([]*pkgCached, 0, 100)
 
 func PkgIDFromPkgPath(path string) PkgID {
        pkgIDMu.Lock()
        defer pkgIDMu.Unlock()
 
-       pkgID, ok := pkgIDFromPkgPathCache[path]
-       if !ok {
-               pkgID = new(PkgID)
-               *pkgID = PkgID{HashBytes([]byte(path))}
-               pkgIDFromPkgPathCache[path] = pkgID
+       // Find the entry if possible, using a slice as
+       // we don't expect too many unique packages and
+       // when small, looking up with a slice is much
+       // cheaper than using a map.
+       for _, cached := range pkgIDFromPkgPathCache {
+               if cached.path == path {
+                       return *cached.id
+               }
+       }
+
+       // Now cache the value.
+       cached := &pkgCached{
+               id:   PkgID{HashBytes([]byte(path))},
+               path: path,
        }
-       return *pkgID
+       pkgIDFromPkgPathCache = append(pkgIDFromPkgPathCache, cached)
+       return *cached.id
 }
```

A map suffices TBH even if memory increases, it is unlikely that 5,000 packages will be imported and thus we can safely assume memory bloat won't be a problem. The merits of the change to me and in experience massively outweight any minuite RAM increases: being able to even start with 8-10 seconds less allows every contributor to work much faster, and even for Gno to preserve compute resources :-)