@odeke-em 2024-12-29
Done, thanks @moul I've added the TODO about using an LRU for the future when the trade-off is needed.